+++ eras/alonzo/impl/testlib/Test/Cardano/Ledger/Alonzo/Arbitrary.hs
@@ -35,6 +35,7 @@ module Test.Cardano.Ledger.Alonzo.Arbitrary (
 
 import Cardano.Ledger.Allegra.Scripts (Timelock)
 import Cardano.Ledger.Alonzo (AlonzoEra, Tx (..))
+import Cardano.Ledger.Alonzo.BlockBody (AlonzoBlockBody (AlonzoBlockBody))
 import Cardano.Ledger.Alonzo.Core
 import Cardano.Ledger.Alonzo.Genesis (AlonzoGenesis (..))
 import Cardano.Ledger.Alonzo.PParams (AlonzoPParams (AlonzoPParams), OrdExUnits (OrdExUnits))
@@ -475,3 +476,13 @@ instance Arbitrary (TransitionConfig AlonzoEra) where
   arbitrary = AlonzoTransitionConfig <$> arbitrary <*> arbitrary
 
 deriving newtype instance Arbitrary (Tx AlonzoEra)
+
+instance
+  ( EraBlockBody era
+  , AlonzoEraTx era
+  , Arbitrary (Tx era)
+  , SafeToHash (TxWits era)
+  ) =>
+  Arbitrary (AlonzoBlockBody era)
+  where
+  arbitrary = AlonzoBlockBody <$> arbitrary