+++ eras/alonzo/test-suite/test/Test/Cardano/Ledger/Alonzo/Golden.hs
@@ -22,7 +22,7 @@ import Cardano.Ledger.Alonzo.TxBody (AlonzoTxOut (..), utxoEntrySize)
 import Cardano.Ledger.BaseTypes (SlotNo (..), StrictMaybe (..), boundRational)
 import Cardano.Ledger.Binary (decCBOR, decodeFullAnnotator)
 import Cardano.Ledger.Binary.Plain as Plain (serialize)
-import Cardano.Ledger.Block (Block (..))
+import Cardano.Ledger.Block (Block (Block))
 import Cardano.Ledger.Coin (Coin (..))
 import Cardano.Ledger.Mary.Value (valueFromList)
 import Cardano.Ledger.Plutus.CostModels (
@@ -296,13 +296,13 @@ goldenMinFee =
               case B16L.decode hex of
                 Left err -> error err
                 Right val -> val
-            txsSeq =
+            blockBody =
               case decodeFullAnnotator (eraProtVerHigh @AlonzoEra) "Block" decCBOR cborBytesBlock of
                 Left err -> error (show err)
-                Right (Block _h txs :: Block (BHeader StandardCrypto) AlonzoEra) -> txs
+                Right (Block _bHeader bBody :: Block (BHeader StandardCrypto) AlonzoEra) -> bBody
             firstTx =
-              case fromTxSeq @AlonzoEra txsSeq of
-                tx :<| _ -> tx
+              case blockBody ^. txSeqBlockBodyL of
+                tx :<| _ -> (tx :: Tx AlonzoEra)
                 Empty -> error "Block doesn't have any transactions"
 
             -- Below are the relevant protocol parameters that were active